requirements.txt
aiogram
environs
psycopg2-binary